refactor: refactor file processor component
