+++ apps/api/src/upload/fileProcessor/processors/index.ts
@@ -1,0 +1,5 @@
+import { PlaylistProcessor } from './playlistProcessor';
+import { StreamHistoryProcessor } from './streamHistoryProcessor';
+import { SimpleStreamHistoryProcessor } from './simpleStreamHistoryProcessor';
+
+export { StreamHistoryProcessor, PlaylistProcessor, SimpleStreamHistoryProcessor };
